Notexactlyawe/GitHub hmac (#1)
* Add base16 and split options

base16 is an option that dictates whether or not the HMAC comes as a
hex string
split is whether the HMAC comes as 'algo=digest'. the alternative is
'digest'
gitignore changes for vim

* Remove b16 flag since handled in digest

Turns out you can use the digest option to do the same thing that the
b16 flag would.

* Add split_digest docs in README

* Add tests + split_hash bug

There was a bug with base64 hashes that contain '=' when using
split_digest. They would always be unauthorized because the hash was
being split on '='. Fixed bug by adding parts: 2 flag that only splits
on the first equals